Split out FluxDensityModel into its own submodule
The FluxDensityModel is now in its own file for better traceability, also
leaving more space to add new flux density models to the file without
cluttering up the already groaning target.py.
