+++ katpoint/test/__init__.py
@@ -9,6 +9,7 @@ import test_antenna
 import test_catalogue
 import test_projection
 import test_timestamp
+import test_flux
 import test_conversion
 import test_pointing
 import test_refraction
@@ -29,6 +30,7 @@ def suite():
     testsuite.addTests(loader.loadTestsFromModule(test_catalogue))
     testsuite.addTests(loader.loadTestsFromModule(test_projection))
     testsuite.addTests(loader.loadTestsFromModule(test_timestamp))
+    testsuite.addTests(loader.loadTestsFromModule(test_flux))
     testsuite.addTests(loader.loadTestsFromModule(test_conversion))
     testsuite.addTests(loader.loadTestsFromModule(test_pointing))
     testsuite.addTests(loader.loadTestsFromModule(test_refraction))
